backend/pyproject.toml
[project]
name = "leaf"
version = "0.1.0"
description = "Leaf API Service"
authors = [
    {name = "Lukas Santing",email = "lukas.santing@gmail.com"}
]
readme = "README.md"

[tool.poetry]
name = "leaf"
version = "0.1.0"
description = "Leaf API Service"
authors = ["Lukas Santing <lukas.santing@gmail.com>"]
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.109.0"
uvicorn = "^0.27.0"
loguru = "^0.7.3"
sqlalchemy = "^2.0.41"
pydantic-settings = "^2.9.1"
pyyaml = "^6.0"
markdownify = "^0.12"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
